Switch multibinding contribution fields to use owned bindings instead of inherited bindings, if both exist.
This has no observable effect (the only change is to generated code that determines which field to reference in a subcomponent), but this ensures that the binding referenced is the one that the BindingGraph expects. This will make a difference when we implement lazy initialization, since then the package proxies will have references based on the BindingGraph, and so the fields should mirror that structure.

-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=96040761